backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
prisma==0.13.1
python-multipart==0.0.9
httpx==0.27.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
apscheduler==3.10.4
